init with swift package manager
