--- LNZTreeView/LNZTreeView.h
@@ -1,19 +1,0 @@
-//
-//  LNZTreeView.h
-//  LNZTreeView
-//
-//  Created by Giuseppe Lanza on 07/11/2017.
-//  Copyright © 2017 Giuseppe Lanza. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for LNZTreeView.
-FOUNDATION_EXPORT double LNZTreeViewVersionNumber;
-
-//! Project version string for LNZTreeView.
-FOUNDATION_EXPORT const unsigned char LNZTreeViewVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <LNZTreeView/PublicHeader.h>
-
-
