--- LNZTreeViewDemo.xcodeproj/xcuserdata/giuseppelanza.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>LNZTreeView.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>1</integer>
-		</dict>
-		<key>LNZTreeViewDemo.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>50D0A40E1FB2215300DE8E02</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>50D0A4411FB22ABF00DE8E02</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>50D0A4491FB22ABF00DE8E02</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
