--- LNZTreeViewDemo/Assets.xcassets/index_folder_indicator.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "index_folder_indicator.pdf"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
